Refactored MARC-related functionality.
